requirements.txt
# Core dependencies
numpy==1.26.4
requests>=2.31.0,<2.32.0
tqdm
wget

# Audio processing
ffmpeg-python>=0.2.0
faiss-cpu==1.7.3
librosa==0.11.0
scipy==1.11.1
soundfile==0.12.1
sox==1.5.0
soxr==0.5.0
noisereduce
pedalboard
stftpitchshift
flatten-dict==0.4.2
torch-dct==0.1.6
auraloss==0.4.0
pesq @ https://github.com/codename0og/codename-essentials/raw/refs/heads/main/pesq-0.0.5-cp310-cp310-win_amd64.whl; sys_platform == "win32"

# EXP
PyWavelets==1.8.0

# Machine learning and deep learning
numba==0.61.0; sys_platform == 'win32'
torch==2.7.1
torchaudio==2.7.1
torchvision==0.22.1
torchcrepe==0.0.23
torchfcpe
torch-optimizer
einop
transformers==4.44.2
nnaudio==0.3.3
ring-attention-pytorch @ https://github.com/codename0og/codename-essentials/raw/refs/heads/main/ring_attention_pytorch-0.5.17-py3-none-any.whl; sys_platform == "win32"
triton-windows; sys_platform == "win32"

# Visualization and UI
matplotlib==3.7.2
tensorboard
gradio==5.23.1

# Miscellaneous utilities
tensorboardX
edge-tts==6.1.9
pypresence
beautifulsoup4
flask
psutil==7.0.0
zstandard